Added GPT Translator Source Files and Responses
The source files used in the original GPT Translator are a combination of .txt and .sysml files. There are ten pairs of these combinations that were added to Chat GPT's knowledge to produce better outputs. The .txt with no suffix  (Control, Translated, JSON) is the actual description of the model/system. The .txt with ModelNameControl.txt was a response generated by prompting Chat GPT-4o with a .jpg image of the SysMLv2 model. The .txt with the suffix Translated is the fine-tuned GPT's response. It used the .txt and .sysml source files to generate responses. The responses are saved in text files so that we can compare them later on in the BERT Score and MAUVE Analysis assessments. The .svg files are not relevant and were originally how I downloaded model images before converting them into .jpg.