Models not in Source Files Added
The five models added here are not in the GPT's knowledge base. They are models outside of the knowledge to test the GPT's ability to translate SysMLv2 model images that it is not familiar with. This was done with all three GPT's: the Control (GPT-4o), Translator GPT, and JSON GPT. The JSON GPT had 20 models in its source files as it used a more efficient way of storing the fine tuned data (GPT's knowledge is capped at 20 files so using JSON allowed me to get double the models in the JSON GPT's knowledge base). The .txt files here follow the same naming convention before as dictated by their suffixes and will be graded based on their BERT and MAUVE scores.