Added BERTScore Files
Includes the demo for a tutorial on how the BERTScore is used with the .txt files referenced in the demo and all 25 responses to every model for each GPT. This includes the Control GPT (GPT-4o), The GPT Translator (source files are .txt and .sysml with 10 models), the JSON GPT (operates the same as the GPT Translator but with JSON sources files and 20 models instead), and the actual textual descriptions of the models. These files are used to grade the results of each GPT based on the actual textual descriptions.