Poster Presented At NEEC Fest
Attached is the poster I showed at the biannual NEEC Fest in Corona, California. The poster gives an overview of how the team is looking to approach Model-Based Systems/Mission Engineering by utilizing AI and LLMs to expedite the digital transformation of knowledge, artifacts, and processes, into data, models, and code. We seek to address the adoption of LLMs to utilize in defense by recognizing both the risks and opportunities Generative AI creates. We hope to understand the limitations and the path to reaching a private LLM fit for defense purposes by the end of the summer. Additionally, by the end of next year, we desire to consolidate our work and create an LLM exemplar that is fit for purpose.